requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
mysql-connector-python==8.2.0
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
Werkzeug==3.0.1
WTForms==3.1.1

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pdf2image==1.16.3
Pillow==10.1.0

# OCR
pytesseract==0.3.10

# Machine Learning
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
joblib==1.3.2

# NLP
nltk==3.8.1

# Utilities
python-dateutil==2.8.2
watchdog==3.0.0
